benerin controller

+++ controllers/book.controller.js
@@ -7,12 +7,18 @@ const path = require('path')
 const fs = require(`fs`)
 
 exports.getAllBook = async(request, response) => {
-    let books = await bookModel.findAll() 
+    let book = await userModel.findAll()
+    if (book.length === 0) {
+        return response.status(400).json({
+            success: false,
+            message: "no book to show",
+        });
+    }
     return response.json({
-        success: true, 
-        data: books,
-        message: `All Book have been loaded`
-    })    
+        success: true,
+        data: book,
+        message: `All books have been loaded`
+    })   
 }
 
 exports.findBook = async (request, response) => {
@@ -20,12 +26,8 @@ exports.findBook = async (request, response) => {
     let books = await bookModel.findAll({ 
         where: {
             [Op.or]: [
-                { BookID: { [Op.substring]: keyword } },
-                { judul: { [Op.substring]: keyword } },
                 { penulis: { [Op.substring]: keyword } },
-                { foto: { [Op.substring]: keyword } },
-                { harga: { [Op.substring]: keyword } },
-                { KategoriID: { [Op.substring]: keyword } }
+                { judul: { [Op.substring]: keyword } }
             ]
         }
     })
@@ -74,12 +76,24 @@ exports.addBook = (request, response) => {
 
 exports.updateBook = (request, response) => {
     upload(request, response, async error => {
-        /** check if there are error when upload */
         if (error) {
             return response.json({ message: error })
         }
 
         let BookID = request.params.BookID
+
+        let getId = await bookModel.findAll({
+            where: {
+                [Op.and]: [{ id: BookID }],
+            },
+        });
+
+        if (getId.length === 0) {
+            return response.status(400).json({
+                success: false,
+                message: "buku dengan id tersebut tidak ada",
+            });
+        }
 
         let dataBook = {
             BookID: request.body.BookID,
@@ -90,27 +104,31 @@ exports.updateBook = (request, response) => {
         }
  
         if (request.file) {
-            /** get selected event's data */
             const selectedBook = await bookModel.findOne({
                 where: { BookID: BookID }
             })
             const oldImage = selectedBook.image
-
-            /** prepare path of old image to delete file */
             const pathImage = path.join(__dirname, `../image`, oldImage)
 
             if (fs.existsSync(pathImage)) {
-                /** delete old image file */
                 fs.unlink(pathImage, error => console.log(error))
             }
-
-            /** add new image filename to event object */
             dataBook.image = request.file.filename    
 
         }
 
-        
-        
+        if (
+            dataBook.Judul === "" ||
+            dataBook.Penulis === "" ||
+            dataBook.Harga === "" ||
+            dataBook.KategoriID === ""
+        ) {
+            return response.status(400).json({
+                success: false,
+                message:
+                    "Harus diisi semua kalau tidak ingin merubah isi dengan value sebelumnya",
+            });
+        }
 
         bookModel.update(dataBook, { where: { id: BookID } })
             .then(result => {
@@ -129,36 +147,36 @@ exports.updateBook = (request, response) => {
 }
 
 exports.deleteBook = async (request, response) => {
-    /** store selected event's ID that will be delete */
     const BookID = request.params.BookID
+    let getId = await userModel.findAll({
+        where: {
+            [Op.and]: [{ id: BookID }],
+        },
+    });   
+    
+    if (getId.length === 0) {
+        return response.status(400).json({
+            success: false,
+            message: "buku dengan id tersebut tidak ada",
+        });
+    }
 
-    /** -- delete image file -- */
-    /** get selected event's data */
-    const book = await bookModel.findOne({ where: { BookID: BookID } })
-    /** get old filename of image file */
+    const book = await bookModel.findOne({ where: { id: BookID } });
     const oldImage = book.foto
-
-    /** prepare path of old image to delete file */
     const pathImage = path.join(__dirname, `../image`, oldImage)
 
-    /** check file existence */
     if (fs.existsSync(pathImage)) {
-        /** delete old image file */
        fs.unlink(pathImage, error => console.log(error))
     }
-    /** -- end of delete image file -- */
 
-    /** execute delete data based on defined id event */
     bookModel.destroy({ where: { BookID: BookID } })
         .then(result => {
-            /** if update's process success */
             return response.json({
                 success: true,
                 message: `Data Book has been deleted`
             })
         })
         .catch(error => {
-            /** if update's process fail */
             return response.json({
                 success: false,
                 message: error.message
