+++ controllers/detailtransaksi.controller.js
@@ -1,10 +1,15 @@
 const { request, response } = require("express")
-
-const bookModel =require(`../models/index`).book
 const detailtransaksiModel = require(`../models/index`).detailtransaksi
 const Op = require(`sequelize`).Op
+
 exports.getAllDetailtransaksi = async(request, response) => {
     let detailtransaksis = await detailtransaksiModel.findAll() 
+    if (detailtransaksis.length === 0) {
+        return response.status(400).json({
+            success: false,
+            message: "nothing user to show",
+        });
+    }
     return response.json({
         success: true, 
         data: detailtransaksis,
@@ -24,11 +29,19 @@ exports.findDetailtransaksi = async (request, response) => {
             ]
         }
     })
-    return response.json({
-         success: true, 
-         data: detailtransaksis,
-         message: `All Detail Transaksi have been loaded`
-    })
+
+    if (!detailtransaksis) {
+        return response.status(400).json({
+            success: false,
+            message: "no details to show",
+        });
+    } else {
+        return response.json({
+            success: true, 
+            data: detailtransaksis,
+            message: `All Detail Transaksi have been loaded`
+       })
+    }
 }
 
 exports.addDetailtransaksi = (request, response) => {
@@ -38,7 +51,6 @@ exports.addDetailtransaksi = (request, response) => {
         Qyt: request.body.Qyt
     }
 
-    
     detailtransaksiModel.create(newDetailtransaksi)
     .then(result => {
         return response.json({
@@ -55,14 +67,38 @@ exports.addDetailtransaksi = (request, response) => {
     })
 }
 
-exports.updateDetailtransaksi = (request, response) => {
+exports.updateDetailtransaksi = async (request, response) => {
+    let DetailTraID = request.params.DetailTraID
+    let getId = await detailtransaksiModel.findAll({
+        where: {
+            [Op.and]: [{ id: DetailTraID }],
+        },
+    });
+
+    if (getId.length === 0) {
+        return response.status(400).json({
+            success: false,
+            message: "Detail dengan id tersebut tidak ada",
+        });
+    }
     let dataDetailtransaksi = {
         TransaksiID: request.body.TransaksiID,
         BukuID: request.body.BukuID,
         Qyt: request.body.Qyt
     }
 
-    let DetailTraID = request.params.DetailTraID
+    if (
+        dataDetailtransaksi.TransaksiID === "" ||
+        dataDetailtransaksi.BukuID === "" ||
+        dataDetailtransaksi.Qyt === "" 
+    ) {
+        return response.status(400).json({
+            success: false,
+            message:
+                "Harus diisi semua kalau tidak ingin merubah isi dengan value sebelumnya",
+        });
+    }
+
  
     detailtransaksiModel.update(dataDetailtransaksi, { where: { id: DetailTraID } })
         .then(result => {
@@ -79,19 +115,31 @@ exports.updateDetailtransaksi = (request, response) => {
         })
 }
 
-exports.deleteDetailtransaksi = (request, response) => {
-    let DetailTraID = request.params.DetailTraID
+exports.deleteDetailtransaksi = async (request, response) => {
+    let DetailTraID = request.params.id
+    let getId = await detailtransaksiModel.findAll({
+        where: {
+            [Op.and]: [{ id: DetailTraID }],
+        },
+    });
+
+    if (getId.length === 0) {
+        return response.status(400).json({
+            success: false,
+            message: "Detail dengan id tersebut tidak ada",
+        });
+    }
      detailtransaksiModel.destroy({ where: { id: DetailTraID } })
         .then(result => {
             return response.json({
                 success: true,
-                message: `Data Detail Transaksi has been updated`
+                message: `Data Detail Transaksi has been deleted`
             })
         })
         .catch(error => {
-            return response.json({
+            return response.status(400).json({
                 success: false,
                 message: error.message
             })
         })
-    }
+}
